Added appointment controller and modified create appointment method
