+++ MedicalAppointmentApp/Controllers/DoctorController.cs
@@ -1,14 +1,18 @@
 ﻿using MediatR;
 using MedicalAppointmentApp.Data;
+using MedicalAppointmentApp.Data.Models;
 using MedicalAppointmentApp.Mediator.Commands;
 using MedicalAppointmentApp.Mediator.Queries;
 using MedicalAppointmentApp.Models;
 using MedicalAppointmentApp.Models.ViewModels;
 using MedicalAppointmentApp.Queries;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using System;
+using System.Security.Claims;
 using System.Threading.Tasks;
+
 
 namespace MedicalAppointmentApp.Controllers
 {
@@ -42,28 +46,6 @@ namespace MedicalAppointmentApp.Controllers
         {
             var doctorsViewModel = await _mediator.Send(new GetDoctors.Query());
             return View(doctorsViewModel);
-        }
-
-        [ApiExplorerSettings(IgnoreApi = true)]
-        [HttpGet("appointment")]
-        public IActionResult CreateAppointmentView(string doctorId, string institutionId)
-        {
-            var appointmentViewModel = new CreateAppointmentModel() { DoctorId = Int32.Parse(doctorId),
-            //reik ir current user id paduot
-            };
-            return View("CreateAppointment", appointmentViewModel);
-        }
-        [HttpPost("appointment/create")]
-        public async Task<IActionResult> CreateAppointment([FromForm] CreateAppointmentModel appointmentModel)
-        {
-            var response = await _mediator.Send(new CreateAppointment.Command
-            {
-                AppointmentModel = appointmentModel
-            });
-            if (!response.Success)
-                Errors(response);
-
-            return RedirectToAction("DoctorList");
         }
 
         [ApiExplorerSettings(IgnoreApi = true)]
