+++ MedicalAppointmentApp/Models/MapperProfiles/WebMappingProfile.cs
@@ -10,6 +10,7 @@ namespace MedicalAppointmentApp.Models.MapperProfiles
             CreateMap<MedicalSpeciality, GetMedicalSpecialtyModel>();
             CreateMap<Doctor, GetDoctorModel>();
             CreateMap<Institution, GetInstitutionModel>();
+            CreateMap<CreateAppointmentModel, Appointment>();
         }
     }
 }
