+++ MedicalAppointmentApp/Views/Appointment/CreateAppointment.cshtml
@@ -4,7 +4,11 @@
 }
 <h1>yehe dude</h1>
 
-<form asp-controller="Doctor" asp-action="CreateAppointment" method="post">
+<form asp-controller="Appointment" asp-action="CreateAppointment" method="post">
+    <div class="form-group" hidden>
+        <label asp-for="ApplicationUserId"></label>
+        <input asp-for="ApplicationUserId" class="form-control" />
+    </div>
     <div class="form-group" hidden>
         <label asp-for="DoctorId"></label>
         <input asp-for="DoctorId" class="form-control" />
