+++ MedicalAppointmentApp/Views/Doctor/Doctor.cshtml
@@ -75,7 +75,7 @@
                                         </form>
                                     </td>
                                     <td>
-                                        <input type="button" class="btn btn-info" value="Create Doctor" onclick="location.href='@Url.Action("CreateAppointmentView", "Doctor", parms)'" />
+                                        <input type="button" class="btn btn-info" value="Create Appointment" onclick="location.href='@Url.Action("CreateAppointmentView", "Appointment", parms)'" />
                                         <span class="font-medium mb-0">@institution.Address</span><br>
                                     </td>
                                 </tr>
